bpf: Switch bpf_map_{area_alloc,area_mmapable_alloc}() to u64 size (bsc#1154353).
suse-commit: de2946c30e9ba5343ef7b419b724e48884b52d9a

+++ kernel/bpf/syscall.c
@@ -167,7 +167,7 @@ void *bpf_map_area_alloc(u64 size, int numa_node)
 	return __bpf_map_area_alloc(size, numa_node, false);
 }
 
-void *bpf_map_area_mmapable_alloc(size_t size, int numa_node)
+void *bpf_map_area_mmapable_alloc(u64 size, int numa_node)
 {
 	return __bpf_map_area_alloc(size, numa_node, true);
 }
